requirements.txt
Flask==2.2.3
GitPython==3.1.31
nltk==3.7
Requests==2.31.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
